testing and new images

+++ include/PerlinGrid.h
@@ -1,3 +1,4 @@
+#include <SDL2/SDL_render.h>
 #include <vector>
 #include "FastNoiseLite.h"
 #include "Vector2D.h"
@@ -7,10 +8,9 @@ class PerlinGrid{
     public:
         PerlinGrid();
         ~PerlinGrid();
-        void init(int width, int height, float increment, int magnitude);
+        void init(int width, int height, float increment, float magnitude);
         void update();
         Vector2D getValueAtPosition(float x, float y);
-       
         FastNoiseLite noise;
         int scale, rows, cols, mag;
         std::vector<std::vector<Vector2D>> grid;
